+++ packages/charts/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @frontify/fondue-charts
 
+## 1.8.3
+
+### Patch Changes
+
+- [#1822](https://github.com/Frontify/fondue/pull/1822) [`3470a6a`](https://github.com/Frontify/fondue/commit/3470a6a9671d39fc40c0939bc48df0e8f49bdeeb) Thanks [@SamuelAlev](https://github.com/SamuelAlev)! - fix: add missing dependencies to `external`
+
+- [#1822](https://github.com/Frontify/fondue/pull/1822) [`3470a6a`](https://github.com/Frontify/fondue/commit/3470a6a9671d39fc40c0939bc48df0e8f49bdeeb) Thanks [@SamuelAlev](https://github.com/SamuelAlev)! - refactor: replace `lodash` by `lodash-es`
+
+## 1.8.2
+
+### Patch Changes
+
+- [#1820](https://github.com/Frontify/fondue/pull/1820) [`65a6163`](https://github.com/Frontify/fondue/commit/65a61630c01b2e0492352df44d1956d902543c45) Thanks [@SamuelAlev](https://github.com/SamuelAlev)! - fix: remove deep import from visx as it causes problems
+
 ## 1.8.1
 
 ### Patch Changes